lib: Add start of client library
This is just a nice client library wrapped around the
existing dbus inteface.  It provides caching and matching
up users with ConsoleKit sessions.

It comes from the GDM codebase.

Right now it uses dbus-glib.  The code could be simplified
dramatically if we switched to gdbus.  We should probably
do that.  # Please enter the commit message for your changes. Lines starting