daemon: ignore extraneous SetAutomaticLogin(false) calls for a user
Right now, if a user is not configured for automatic login, and
SetAutomaticLogin(false) is called for that user then we disable
automatic login system wide even if someone else is configured
for automatic login.

This commit changes the semantics of the call such that
SetAutomaticLogin(false) is a no op for users where automatic login
isn't already enabled.

Spotted by Vincent Untz <vuntz@gnome.org>

https://bugs.freedesktop.org/show_bug.cgi?id=39402