lib: change loop iteration idiom
Some of the recent code for iterating through
the results of dbus code uses i when isn't an
integer.  I find that really hard to wrap my head
around, and it appears to be causing crashes.

(I believe the loop condition should be *i not i)

This commit changes things up to be, in my opinion,
clearer.

https://bugs.freedesktop.org/show_bug.cgi?id=46170