ActUserManager: ensure the manager is alive across async calls
Under current design and usage patterns, the library user is not
expected to keep the manager around, if only interested in signals
from ActUsers. Therefore additional references must be taken to
ensure that the object is still valid at the end of an async call,
and signals must be prevented from firing on finalized objects.

https://bugs.freedesktop.org/show_bug.cgi?id=46797