lib: Add missing return in libaccountsservice
In _get_x11_display_for_new_systemd_session there is a missing return.

I found out that running gnome-shell from a remote box via ssh and
setting DISPLAY to :0 the systemd session type is tty. Also the uid
systemd is 0 for my user jhbuild gnome session run from debian gdm3
(ie not based on gnome-shell).
This triggers this error path which miss a "return" instruction.

I have been unable to find out if session type and uid are the same
for the same gdm session with gnome-shell started from gnome-session
though at least it also fixes the segfault there too.

This error code path is not triggered when starting gnome-session with
gnome-shell from a console + xinit inside a "jhbuild shell".
So this might be a weird use case but good to fix nevertheless.

https://bugs.freedesktop.org/show_bug.cgi?id=48478