daemon: add all users from cachedir to ListCachedUsers result
The cachedir /var/lib/AccountsService/users/ contains a list of
files named after user accounts.  These files contain metadata
about the user.

This commit changes the code to consider the presence of one of
these cache files as evidence of a user account, and uses that
information to augment the list of users returned by ListCachedUsers.

This prevents network-login users from "falling off the map" if the
wtmp file is rotated.

It's also a prerequisite for adding support for an explicit
CacheUser() call. That call will be introduced in a subsequent
commit.

https://bugs.freedesktop.org/show_bug.cgi?id=50770