daemon: uncache user before deleting user
When deleting a user account it's important to clean up our
cache file on the user even if userdel fails.  userdel only
works on local users, and we want DeleteUser() when called
on remote users to behave something like UncacheUser().

This commit rearranges the order in which we delete the users
cache files versus delete the user, such that the cache files
are always deleted.

https://bugs.freedesktop.org/show_bug.cgi?id=50770